requirements.txt
# Dépendances principales pour sync_woocommerce_odoo
requests
odoorpc
python-dotenv
woocommerce
